Let Ruby do the hard work for us.
I was resistent to metaprogramming and passing the enablement hashes
straight through to the class, but this lets us delete a lot of code and
supply defaults and do a lot of fun stuff with no work. Also added a
better error if you do specify an enablement method that we don't know
about.