Add inter-feature dependency.
Yes, you could do a circular dependency. It seems like overkill, but I
set this up in a real app and manually mentioned in three different
places that one feature wouldn't work until another one was enabled. Now
we can show that *and* have it affect the enablement status of the
feature.