Add #permanently_disabled support.
This was intended as a way to turn off a feature you never want on for
any reason. It works, but the truth is that there are some odd situation
where it just doesn't work. If you have the enablement check at the
class level and the class autoloads and/or you have a forking situation
like happens with selenium, it's not working so hot.